feat: added get all user email function
